Moves sha512t_h0 tests out of unit tests and into integration tests. Added a note about this to QUALITY_AND_STYLE.md.
