+++ crypto/sha2/src/lib.rs
@@ -1,10 +1,13 @@
 //! Implements SHA2 as per NIST FIPS 180-4.
 //!
+//! This crate provides the following primitives:
+//!
+//! * SHA2 [`Hash`] functions.
+//!
 //! # Examples
 //! ## Hash
 //! Hash functionality is accessed via the [`bouncycastle_core::traits::Hash`] trait,
-//! which is implemented by [`SHA224`], [`SHA256`], [`SHA384`], [`SHA512`], [`SHA512_224`] and
-//! [`SHA512_256`].
+//! which is implemented by all the SHA2 primitives.
 //!
 //! The simplest usage is via the static functions.
 //! ```
@@ -16,7 +19,7 @@
 //! ```
 //!
 //! More advanced usage will require creating a SHA2 object to hold state between successive calls,
-//! for example if input is received in chunks and not all available at the same time:
+//! for example, if input is received in chunks and not all available at the same time:
 //!
 //! ```
 //! use bouncycastle_sha2 as sha2;
@@ -35,10 +38,11 @@
 //! let output: Vec<u8> = sha2.do_final();
 //! ```
 //!
+//! ## Partial byte
 //! It is also possible to provide input where the final byte contains fewer than 8 bits of data
-//! (a bit-oriented message, FIPS 180-4 s. 5.1). The partial byte is taken as it arrives in the final
-//! octet of an ASN.1 BIT STRING: the message bits are its most significant bits, leading bit first, and
-//! the low "unused" bits are ignored. The following hashes 16 bytes plus the 3 bits `101`:
+//! (a bit-oriented message, FIPS 180-4 s. 5.1). The partial byte is taken as the most significant bits,
+//! leading bit first, and the low "unused" bits are ignored. The following hashes 16 bytes plus the
+//! 3 padding bits `101`:
 //! ```
 //! use bouncycastle_core::traits::Hash;
 //! use bouncycastle_sha2 as sha2;
@@ -48,68 +52,6 @@
 //! sha2.do_update(&data[..16]);
 //! let output: Vec<u8> = sha2.do_final_partial_bits(data[16], 3).expect("num_partial_bits is in 0..=7");
 //! ```
-//!
-//! # SHA-512/t
-//!
-//! FIPS 180-4 s. 5.3.6 defines SHA-512/t, a family of hash functions that run SHA-512 with a
-//! t-specific initial hash value and truncate the result to t bits. The family is exposed as the
-//! generic [`SHA512t`]; its initial hash value is derived at compile time by the spec's "SHA-512/t
-//! IV Generation Function". Only the two truncations that FIPS 180-4 approves, `t = 224` and
-//! `t = 256`, are instantiable, as [`SHA512_224`] and [`SHA512_256`]; any other `t` fails to
-//! compile.
-//!
-//! ```
-//! use bouncycastle_core::traits::Hash;
-//! use bouncycastle_sha2 as sha2;
-//!
-//! let output: Vec<u8> = sha2::SHA512_256::new().hash(b"Hello, world!");
-//! assert_eq!(output.len(), 32);
-//!
-//! // `SHA512_256` is an alias for `SHA512t<256>`.
-//! let same: Vec<u8> = sha2::SHA512t::<256>::new().hash(b"Hello, world!");
-//! assert_eq!(output, same);
-//! ```
-//!
-//! A truncation that FIPS 180-4 does not approve is rejected by the compiler:
-//!
-//! ```compile_fail
-//! use bouncycastle_core::traits::Hash;
-//! use bouncycastle_sha2 as sha2;
-//!
-//! let output: Vec<u8> = sha2::SHA512t::<200>::new().hash(b"Hello, world!");
-//! ```
-//!
-//! # Memory Usage
-//!
-//! No heap memory is used by the algorithms themselves; the `Vec<u8>`-returning convenience methods
-//! allocate only the output buffer, and the `*_out` variants allocate nothing.
-//!
-//! | Object                                                   | Size (bytes) |
-//! |----------------------------------------------------------|--------------|
-//! | `SHA224`, `SHA256`                                       | 112          |
-//! | `SHA384`, `SHA512`, `SHA512_224`, `SHA512_256`           | 208          |
-//! | Suspended `SHA224`/`SHA256` state                        | 108          |
-//! | Suspended `SHA384`/`SHA512`/`SHA512_224`/`SHA512_256` state | 204       |
-//!
-//! The object holds the 8-word chaining value plus one block of buffered input. The compression
-//! function additionally uses a 64-word (SHA-256 family, 256 bytes) or 80-word (SHA-512 family,
-//! 640 bytes) message schedule on the stack for the duration of a call.
-//!
-//! # Security Considerations
-//!
-//! * SHA-224/256/384/512 offer 112/128/192/256 bits of collision resistance respectively;
-//!   SHA-512/224 and SHA-512/256 offer 112 and 128 bits.
-//! * SHA-2 is a Merkle–Damgård construction and is therefore subject to length-extension:
-//!   `H(k || m)` is not a secure MAC. Use HMAC (`bouncycastle-hmac`) for keyed hashing.
-//! * SHA-224, SHA-384, SHA-512/224 and SHA-512/256 are truncations of SHA-256 or SHA-512 with
-//!   distinct initial values, and are not vulnerable to length extension in the same direct way, but
-//!   should still not be used as `H(k || m)` MACs.
-//! * The chaining value and input buffer are held in [`bouncycastle_utils::secret::Secret`] and
-//!   zeroized on drop. Transient copies (working variables and message schedule) in registers/stack
-//!   locals during compression are not zeroized.
-//! * The implementation contains no data-dependent branches or table lookups.
-//! * Messages up to 2^64 bytes are supported (FIPS 180-4 permits 2^64 bits for SHA-224/256 and
-//!   2^128 bits for SHA-384/512 and SHA-512/t; the SHA-512 family limit here is 2^67 bits).
 //!
 //! # Suspending and resuming execution
 //!
@@ -141,6 +83,31 @@
 //! sha2_resumed.do_update(msg_part2);
 //! let h: Vec<u8> = sha2_resumed.do_final();
 //! ```
+//!
+//! # Memory Usage
+//!
+//! | Object                                                   | Size (bytes) |
+//! |----------------------------------------------------------|--------------|
+//! | `SHA224`, `SHA256`                                       | 112          |
+//! | `SHA384`, `SHA512`, `SHA512_224`, `SHA512_256`           | 208          |
+//! | Suspended `SHA224`/`SHA256` state                        | 108          |
+//! | Suspended `SHA384`/`SHA512`/`SHA512_224`/`SHA512_256` state | 204       |
+//!
+//! # Security Considerations
+//!
+//! * SHA-224/256/384/512 offer 112/128/192/256 bits of collision resistance respectively;
+//!   SHA-512/224 and SHA-512/256 offer 112 and 128 bits (SP 800-107r1, Table 1 (§4.2)).
+//! * SHA-2 is a Merkle–Damgård construction and is therefore subject to length-extension:
+//!   `H(k || m)` is not a secure MAC. Use HMAC (`bouncycastle-hmac`) for keyed hashing.
+//! * SHA-224, SHA-384, SHA-512/224 and SHA-512/256 are truncations of SHA-256 or SHA-512 with
+//!   distinct initial values, and are not vulnerable to length extension in the same direct way, but
+//!   should still not be used as `H(k || m)` MACs.
+//! * The chaining value and input buffer are held in [`bouncycastle_utils::secret::Secret`] and
+//!   zeroized on drop. Transient copies (working variables and message schedule) in registers/stack
+//!   locals during compression are not zeroized.
+//! * The implementation contains no data-dependent branches or table lookups.
+//! * Messages up to 2^64 bytes are supported (FIPS 180-4 permits 2^64 bits for SHA-224/256 and
+//!   2^128 bits for SHA-384/512 and SHA-512/t; the SHA-512 family limit here is 2^67 bits).
 
 #![forbid(unsafe_code)]
 #![forbid(missing_docs)]
@@ -158,6 +125,7 @@ use bouncycastle_core::traits::{Algorithm, AlgorithmOID, HashAlgParams, Security
 /*** Imports needed for docs ***/
 #[allow(unused_imports)]
 use bouncycastle_core::traits::{Hash, Suspendable};
+/*** end of doc-only imports ***/
 
 /*** String constants ***/
 /// Algorithm name string for SHA224, as used by the factories and CLI.
@@ -327,37 +295,6 @@ impl Sha512Family for SHA512Params {
 #[derive(Clone)]
 pub struct SHA512tParams<const T: usize>;
 
-/// FIPS 180-4 s. 5.3.6.1: the eight 64-bit words H(0) shall consist of for SHA-512/224, "obtained
-/// by executing the SHA-512/t IV Generation Function with t = 224".
-const SHA512_224_H0: [u64; 8] = [
-    0x8C3D37C819544DA2, 0x73E1996689DCD4D6, 0x1DFAB7AE32FF9C82, 0x679DD514582F9FCF,
-    0x0F6D2B697BD44DA8, 0x77E36F7304C48942, 0x3F9D85A86A1D36C8, 0x1112E6AD91D692A1,
-];
-
-/// FIPS 180-4 s. 5.3.6.2: the eight 64-bit words H(0) shall consist of for SHA-512/256, "obtained
-/// by executing the SHA-512/t IV Generation Function with t = 256".
-const SHA512_256_H0: [u64; 8] = [
-    0x22312194FC2BF72C, 0x9F555FA3C84C64C2, 0x2393B86B6F53B151, 0x963877195940EABD,
-    0x96283EE2A88EFFE3, 0xBE5E1E2553863992, 0x2B0199FC2C85B8AA, 0x0EB72DDC81C52CA2,
-];
-
-/// `const`-evaluable `a == b` for the H(0) arrays (array `PartialEq` is not `const`).
-const fn h0_eq(a: &[u64; 8], b: &[u64; 8]) -> bool {
-    let mut i = 0;
-    while i < 8 {
-        if a[i] != b[i] {
-            return false;
-        }
-        i += 1;
-    }
-    true
-}
-
-// The IV Generation Function (s. 5.3.6) must reproduce the words listed in s. 5.3.6.1 and
-// s. 5.3.6.2. Checked at compile time, so a wrong H(0) can never reach a build.
-const _: () = assert!(h0_eq(&sha512t_h0(224), &SHA512_224_H0), "FIPS 180-4 s. 5.3.6.1");
-const _: () = assert!(h0_eq(&sha512t_h0(256), &SHA512_256_H0), "FIPS 180-4 s. 5.3.6.2");
-
 /*** SHA512/224 ***/
 impl Algorithm for SHA512tParams<224> {
     const ALG_NAME: &'static str = SHA512_224_NAME;
@@ -375,7 +312,8 @@ impl AlgorithmOID for SHA512_224 {
 }
 impl SHA2Params for SHA512tParams<224> {}
 impl Sha512Family for SHA512tParams<224> {
-    // FIPS 180-4 s. 6.6 exception 1: H(0) as specified in s. 5.3.6.1 (checked against it above).
+    // FIPS 180-4 s. 6.6 exception 1: H(0) as specified in s. 5.3.6.1 (pinned against the words
+    // listed there by tests/sha512t_h0_tests.rs).
     const H0: [u64; 8] = sha512t_h0(224);
 }
 
@@ -396,39 +334,9 @@ impl AlgorithmOID for SHA512_256 {
 }
 impl SHA2Params for SHA512tParams<256> {}
 impl Sha512Family for SHA512tParams<256> {
-    // FIPS 180-4 s. 6.7 exception 1: H(0) as specified in s. 5.3.6.2 (checked against it above).
+    // FIPS 180-4 s. 6.7 exception 1: H(0) as specified in s. 5.3.6.2 (pinned against the words
+    // listed there by tests/sha512t_h0_tests.rs).
     const H0: [u64; 8] = sha512t_h0(256);
-}
-
-/// `h0_eq` and `sha512t_h0` are otherwise only evaluated inside `const` assertions, which
-/// `cargo mutants` cannot see fail (a mutant that makes `h0_eq` always true just makes the assertions
-/// vacuous), so they are exercised at runtime here as well.
-#[cfg(test)]
-mod const_helper_tests {
-    use super::*;
-
-    #[test]
-    fn h0_eq_detects_a_difference_in_any_word() {
-        assert!(h0_eq(&SHA512_224_H0, &SHA512_224_H0));
-        assert!(!h0_eq(&SHA512_224_H0, &SHA512_256_H0));
-        for i in 0..8 {
-            let mut h = SHA512_256_H0;
-            h[i] ^= 1;
-            assert!(!h0_eq(&h, &SHA512_256_H0), "word {i}");
-        }
-    }
-
-    /// FIPS 180-4 s. 5.3.6.1 / s. 5.3.6.2: the IV Generation Function reproduces the listed words.
-    #[test]
-    fn sha512t_h0_matches_the_listed_words() {
-        assert_eq!(sha512t_h0(224), SHA512_224_H0);
-        assert_eq!(sha512t_h0(256), SHA512_256_H0);
-        assert_eq!(<SHA512tParams<224> as Sha512Family>::H0, SHA512_224_H0);
-        assert_eq!(<SHA512tParams<256> as Sha512Family>::H0, SHA512_256_H0);
-        // FIPS 180-4 s. 5.3.6: the two-digit and one-digit t paths of the message formatting.
-        assert_ne!(sha512t_h0(8), sha512t_h0(80));
-        assert_ne!(sha512t_h0(80), sha512t_h0(224));
-    }
 }
 
 pub use sha256::SUSPENDED_SHA256_STATE_LEN;
